server/requirement.txt
dotenv
google-genai
flask
flask_cors